spl_pod

Module bytemuck

wrappers for bytemuck functions

Functions

pod_bytes_of

Convert a Pod into a slice of bytes (zero copy)

pod_from_bytes

Convert a slice of bytes into a Pod (zero copy)

pod_from_bytes_mut

Convert a slice of bytes into a mutable Pod (zero copy)

pod_get_packed_len

On-chain size of a Pod type

pod_maybe_from_bytes

Maybe convert a slice of bytes into a Pod (zero copy)

pod_slice_from_bytes

Convert a slice of bytes into a Pod slice (zero copy)

pod_slice_from_bytes_mut

Convert a slice of bytes into a mutable Pod slice (zero copy)

pod_slice_to_bytes

Convert a Pod slice into a single slice of bytes