spl_pod::option

Trait Nullable

pub trait Nullable:
    PartialEq
    + Pod
    + Sized {
    const NONE: Self;

    // Provided methods
    fn is_none(&self) -> bool { ... }
    fn is_some(&self) -> bool { ... }
}

Trait for types that can be None.

This trait is used to indicate that a type can be None according to a specific value.

Required Associated Constants

const NONE: Self

Value that represents None for the type.

Provided Methods

fn is_none(&self) -> bool

Indicates whether the value is None or not.

fn is_some(&self) -> bool

Indicates whether the value is Some`` value of type T`` or not.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.