Crate spl_pod

Crate containing Pod types and bytemuck utilities used in SPL

Re-exports

pub use solana_decode_error;

pub use solana_msg;

pub use solana_program_error;

pub use solana_program_option;

pub use solana_pubkey;

Modules

bytemuck

wrappers for bytemuck functions

error

Error types

option

Generic Option that can be used as a Pod for types that can have a designated None value.

optional_keys

Optional pubkeys that can be used a Pods

primitives

primitive types that can be used in Pods

slice

Special types for working with slices of Pods

Macros

impl_int_conversion

Simple macro for implementing conversion functions between Pod* integers and standard integers.